requirements.txt
Jinja2==3.1.5
PyYAML==6.0.2
rich==13.7.1